Allow subclasses to instantiate their own specialized implementations (#96)
* Allow subclasses to instantiate their own specialized implementations

The emphasis of this commit is to be able to extend to provide our own
custom ResourceAdjuster, along with custom attributes in the related Ant
tasks.

We have replaced the direct instantiation of key classes via the new
operator with calls to protected methods that create the instances.  We
have made this change for Ant tasks and ResourceAdjuster-related
classes.

A couple classes that were package-private are now accessible for
sub-classing.

Furthermore, we have removed "final" from any class that may now have
customized implementations.

Within these classes, we have also exposed some private fields as
"protected final".

* More fields converted from private to protected

* New method filterContentImpl that allows subclasses to extend

the filtering behavior for a given resource and Adjust section.